Moved testTimerNotInitialize and testTimerInitialize test cases to TestAbfsClient class
